feat: precios de costo, árbol de categorías, paginación y UX improvements
- Ingrediente: agrega precio_costo y stock_minimo con migración automática al startup
- Servicio de ingrediente: actualiza precio de productos al cambiar costo (mantiene margen)
- Seed: sincroniza categorías e ingredientes de productos existentes, agrega precios de costo
- Categorías: vista árbol con conectores CSS, filtrado por nodo hoja en productos
- NumberInput: componente compacto con botones −/+ para inputs de cantidad y stock
- Modal: soporte de z-index configurable para modales anidados
- useUnidadesMedida: migra de fetch crudo a apiClient con auth por cookies
- Admin forms: refactor de ingredientes en 2 columnas, precio libre por teclado
- Store: filtro por ingrediente en HomePage, WebSocket URL fix

+++ backend/models/producto.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timezone
+from decimal import Decimal
 from typing import TYPE_CHECKING, Optional
 from sqlmodel import Field, Relationship, SQLModel, Column, String, BigInteger, Text, Numeric, Boolean, DateTime, ForeignKey, JSON
 from sqlalchemy import CheckConstraint
@@ -34,7 +35,7 @@ class Producto(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     nombre: str = Field(sa_column=Column(String(150), nullable=False))
     descripcion: Optional[str] = Field(default=None, sa_column=Column(Text))
-    precio_base: float = Field(sa_column=Column(Numeric(10, 2), nullable=False))
+    precio_base: Decimal = Field(sa_column=Column(Numeric(10, 2), nullable=False))
     imagenes_url: list[str] = Field(default_factory=list, sa_column=Column(JSON))
     unidad_venta_id: Optional[int] = Field(default=None, foreign_key="unidades_medida.id")
     tiempo_prep_min: Optional[int] = Field(default=None, sa_column=Column(BigInteger))
